+++ lib/presentation/blocs/permit/upload/upload_state.dart
